refactoring

+++ backend/routers/router_messages.py
@@ -7,6 +7,7 @@ from sqlalchemy.orm import Session
 import backend.constants as const
 from backend.db import db_messages
 from backend.db.connection import get_session
+from backend.models.enums import MessageTypeEnum
 from backend.models.model_messages import MessageSchema
 
 router = APIRouter(
@@ -22,8 +23,8 @@ def get_messages(
     session: Session = Depends(get_session)
 ):
     """Вывод всех сообщений"""
-    messages_tuple = db_messages.get_messages(session, offset, limit)
-    return messages_tuple
+    messages = db_messages.get_messages(session, offset, limit)
+    return messages
 
 
 @router.get("/{uid_message}", response_model=MessageSchema)
@@ -38,7 +39,33 @@ def get_message(uid_message: uuid.UUID, session: Session = Depends(get_session))
 @router.post("/", response_model=MessageSchema, status_code=status.HTTP_201_CREATED)
 def create_message(message_schema_item: MessageSchema, session: Session = Depends(get_session)):
     """Создание сообщения"""
-    # TODO скрывать лишние поля, типа как response_model_exclude={"sent", "date_create"}
+    if message_schema_item.id_message_type == MessageTypeEnum.PRIVATE \
+            and not message_schema_item.id_pager:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Для личных сообщений нужно обязательно указать абонентский номер (id_pager)"
+        )
+
+    if message_schema_item.id_message_type == MessageTypeEnum.GROUP \
+            and not message_schema_item.id_group_type:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail=(
+                "Для групповых сообщений нужно обязательно указать "
+                "тип группового сообщения (id_group_type)"
+            )
+        )
+
+    if message_schema_item.id_message_type == MessageTypeEnum.MAILDROP \
+            and not message_schema_item.id_maildrop_type:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail=(
+                "Для maildrop сообщений нужно обязательно указать "
+                "тип новостного сообщения (id_maildrop_type)"
+            )
+        )
+
     message_item = db_messages.create_message(session, message_schema_item)
     if not message_item:
         raise HTTPException(
@@ -54,6 +81,12 @@ def delete_message(uid_message: uuid.UUID, session: Session = Depends(get_sessio
     message_item = db_messages.get_message(session, uid_message)
     if not message_item:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Сообщение не найдено")
+
+    if message_item.sent:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Сообщение уже было отправлено, его нельзя удалить"
+        )
 
     result = db_messages.delete_message(session, uid_message)
     if not result:
